+++ Sources/SmileID/Classes/BiometricKYC/OrchestratedBiometricKycViewModel.swift
@@ -23,6 +23,10 @@ internal class OrchestratedBiometricKycViewModel: ObservableObject {
 
     // MARK: - UI Properties
 
+    /// we use `errorMessageRes` to map to the actual code to the stringRes to allow localization,
+    /// and use `errorMessage` to show the actual platform error message that we show if
+    /// `errorMessageRes` is not set by the partner
+    @Published var errorMessageRes: String?
     @Published var errorMessage: String?
     @Published @MainActor private(set) var step: BiometricKycStep = .selfie
 
@@ -160,13 +164,16 @@ internal class OrchestratedBiometricKycViewModel: ObservableObject {
                 if SmileID.allowOfflineMode, LocalStorage.isNetworkFailure(error: error) {
                     didSubmitBiometricJob = true
                     DispatchQueue.main.async {
-                        self.errorMessage = "Offline.Message"
+                        self.errorMessageRes = "Offline.Message"
                         self.step = .processing(.success)
                     }
                 } else {
                     didSubmitBiometricJob = false
                     print("Error submitting job: \(error)")
+                    let (errorMessageRes, errorMessage) = toErrorMessage(error: error)
                     self.error = error
+                    self.errorMessageRes = errorMessageRes
+                    self.errorMessage = errorMessage
                     DispatchQueue.main.async { self.step = .processing(.error) }
                 }
             } catch {