+++ Sources/SmileID/Classes/SelfieCapture/SelfieViewModel.swift
@@ -45,6 +45,10 @@ public class SelfieViewModel: ObservableObject, ARKitSmileDelegate {
 
     // UI Properties
     @Published var directive: String = "Instructions.Start"
+    /// we use `errorMessageRes` to map to the actual code to the stringRes to allow localization,
+    /// and use `errorMessage` to show the actual platform error message that we show if
+    /// `errorMessageRes` is not set by the partner
+    @Published var errorMessageRes: String?
     @Published var errorMessage: String?
     @Published var processingState: ProcessingState?
     @Published var selfieToConfirm: Data?
@@ -400,12 +404,15 @@ public class SelfieViewModel: ObservableObject, ARKitSmileDelegate {
                 }
                 if SmileID.allowOfflineMode, LocalStorage.isNetworkFailure(error: error) {
                     DispatchQueue.main.async {
-                        self.errorMessage = "Offline.Message"
+                        self.errorMessageRes = "Offline.Message"
                         self.processingState = .success
                     }
                 } else {
                     print("Error submitting job: \(error)")
+                    let (errorMessageRes, errorMessage) = toErrorMessage(error: error)
                     self.error = error
+                    self.errorMessageRes = errorMessageRes
+                    self.errorMessage = errorMessage
                     DispatchQueue.main.async { self.processingState = .error }
                 }
             } catch {